python-st3215/requirements.txt
pyserial
paho-mqtt
psutil
